remove preprocess states, move aug memo to memo file
