+++ rasa_core/policies/__init__.py
@@ -7,7 +7,7 @@ from rasa_core.policies.policy import Policy
 from rasa_core.policies.ensemble import SimplePolicyEnsemble, PolicyEnsemble
 from rasa_core.policies.fallback import FallbackPolicy
 from rasa_core.policies.keras_policy import KerasPolicy
-from rasa_core.policies.memoization import MemoizationPolicy
-from rasa_core.policies.augmented_memoization import AugmentedMemoizationPolicy
+from rasa_core.policies.memoization import \
+    MemoizationPolicy, AugmentedMemoizationPolicy
 from rasa_core.policies.online_trainer import OnlinePolicyEnsemble
 from rasa_core.policies.sklearn_policy import SklearnPolicy
